requirements.txt
selenium==4.1.0
pandas==1.3.3
python-dotenv